[프로모션] 1차쿠폰 계산 - 조건 추가(최소구매금액..)

+++ src/main/java/com/plateer/ec1/promotion/vo/Product.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Builder
 public class Product {
     private String productNo; // 상품번호
-    private int productAmt; // 가격
+    private long productAmt; // 가격
     private int productCnt; // 수량
     private String productItemNo; // 단품번호
 }
